Refactor stats package to use proto.

+++ pkg/logql/stats/grpc_test.go
@@ -81,7 +81,7 @@ func TestCollectTrailer(t *testing.T) {
 		t.Fatal(err)
 	}
 	res := decodeTrailers(ctx)
-	require.Equal(t, 2, res.TotalReached)
+	require.Equal(t, int32(2), res.TotalReached)
 	require.Equal(t, int64(2), res.TotalChunksMatched)
 	require.Equal(t, int64(4), res.TotalBatches)
 	require.Equal(t, int64(6), res.TotalLinesSent)
